small tweaks and refactors
