+++ src/apollo/apollo.ts
@@ -1,6 +1,9 @@
+// Vendor
 import { ApolloClient } from "apollo-client";
 import { InMemoryCache } from "apollo-cache-inmemory";
-import { resolvers } from "../resolvers/index";
+
+// Internal
+import { resolvers } from "./resolvers/index";
 import { typeDefs } from "../types/graphql";
 import { GET_TODOLIST } from "../types/graphql";
